services/dispatch_service/requirements.txt
aiosqlite==0.20.0
anyio==4.3.0
pytest-anyio==0.0.0
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
alembic==1.13.1
pydantic[email]==2.7.1
pydantic-settings==2.2.1
python-jose[cryptography]==3.3.0
aio-pika==9.4.1
redis[asyncio]==5.0.4
httpx==0.27.0
python-dotenv==1.0.1
pytest==8.2.0
pytest-asyncio==0.23.6